kNN seem to work well
